Removed the EventLoop class
The EventLoop was an interesting concept: the combination of an
EventQueue and a Thread. The idea was that the EventLoop would provide
a convenient coupling of these two concepts for use in module
boundaries, and the EventLoop could abstract out some of the
complexities with running an event queue in a thread.

However, with the addition of the chain function, event queues can be
easily composed without threads or indirect references to event queues.
Threads can still be spawned dynamically in default-constructors,
although the overhead is much more explicit and tangible.

Additionally, it turned out there weren't that many complexities with
running an event queue in a thread.

There were, surprisingly, several problems with just passing the
EventQueue::dispatch function to mbed's Thread constructor.

- Split EventQueue::dispatch into overloaded functions
- Remove problematic template overloads in Thread constructor
- Exposed break_dispatch
- Renamed get_tick to tick to match the underlying C api
- Renamed DEFAULT_QUEUE_SIZE to EVENTS_QUEUE_SIZE to avoid name
  possible name collisions