Increased type-infering functions to all possible event permutations
With the stricter type inference on the convenience event function,
it is trivial to determine the quantity of parameters present
in a function type.

With the arity of the function type, we can determine which Event class
must be generated entirely based on type inference. This is a huge
benefit to the value of the convenience event functions and removes
any lingering limitations on the event function usage.