Removed problematic event overloads for function-objects
Unforunately, it is very difficult to infer the parameters of a
function-object generically in C++03 without any additional type
information. The current implementation fails to due so, and the
compiler simply bails with "unable to deduce template parameter".

Rather than leaving the user with a small novel of error messages,
this patch removes the problematic event overloads, leaving only
event overloads for simple pointer types. As a side effect, the
limited sfinae in mbed-events is no longer needed.